+++ spec/octokit/client/apps_spec.rb
@@ -154,19 +154,4 @@ describe Octokit::Client::Apps do
       end # with installed repository on installation
     end # with repository
   end # with app installation
-
-  private
-
-  def new_jwt_token
-    private_pem = File.read(test_github_integration_pem_key)
-    private_key = OpenSSL::PKey::RSA.new(private_pem)
-
-    payload = {}.tap do |opts|
-      opts[:iat] = Time.now.to_i           # Issued at time.
-      opts[:exp] = opts[:iat] + 600        # JWT expiration time is 10 minutes from issued time.
-      opts[:iss] = test_github_integration # Integration's GitHub identifier.
-    end
-
-    JWT.encode(payload, private_key, 'RS256')
-  end
 end
