+++ spec/octokit/client/marketplace_spec.rb
@@ -1,6 +1,6 @@
 require 'helper'
 
-describe Octokit::Client::MarketplaceListings do
+describe Octokit::Client::Marketplace do
   before(:each) do
     Octokit.reset!
     @client     = oauth_client
@@ -35,19 +35,4 @@ describe Octokit::Client::MarketplaceListings do
       assert_requested :get, github_url("/marketplace_listing/accounts/1")
     end
   end # .plan_for_account
-
-  private
-
-  def new_jwt_token
-    private_pem = File.read(test_github_integration_pem_key)
-    private_key = OpenSSL::PKey::RSA.new(private_pem)
-
-    payload = {}.tap do |opts|
-      opts[:iat] = Time.now.to_i           # Issued at time.
-      opts[:exp] = opts[:iat] + 600        # JWT expiration time is 10 minutes from issued time.
-      opts[:iss] = test_github_integration # Integration's GitHub identifier.
-    end
-
-    JWT.encode(payload, private_key, 'RS256')
-  end
 end
